requirements.txt
# Framework
Django==4.0

# Tools to create interfaces over REST sta
djangorestframework==3.13.1

# Endpoints Documentation
drf-yasg==1.20.0

# search filters with GET params
django-filter==21.1


